Move Converter to separate file + adapt examples
